requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-docx